requirements.txt
pdf2image==1.16.3
requests==2.31.0
Pillow==10.3.0
